+++ packages/colony-js-client/src/TokenClient/index.js
@@ -54,7 +54,7 @@ export default class TokenClient extends ContractClient {
     },
     TokenClient,
     {
-      contract: 'erc20.sol',
+      contract: 'base.sol',
       // eslint-disable-next-line max-len
       contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
@@ -73,9 +73,9 @@ export default class TokenClient extends ContractClient {
     },
     TokenClient,
     {
-      contract: 'token.sol',
+      contract: 'Token.sol',
       // eslint-disable-next-line max-len
-      contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
+      contractPath: 'https://github.com/JoinColony/colonyToken/blob/7cc7d6b5bf3e94e6d97cd65583e3da38a994753f/contracts',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
     },
   >;
@@ -93,7 +93,7 @@ export default class TokenClient extends ContractClient {
     TokenClient,
     {
       function: 'allowance',
-      contract: 'erc20.sol',
+      contract: 'base.sol',
       // eslint-disable-next-line max-len
       contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
@@ -112,7 +112,7 @@ export default class TokenClient extends ContractClient {
     TokenClient,
     {
       function: 'balanceOf',
-      contract: 'erc20.sol',
+      contract: 'base.sol',
       // eslint-disable-next-line max-len
       contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
@@ -132,7 +132,7 @@ export default class TokenClient extends ContractClient {
     {
       contract: 'Token.sol',
       // eslint-disable-next-line max-len
-      contractPath: 'https://github.com/JoinColony/colonyToken/blob/2dc3136d71c0af9adb9b05e187885dd9dcae94c8/contracts',
+      contractPath: 'https://github.com/JoinColony/colonyToken/blob/7cc7d6b5bf3e94e6d97cd65583e3da38a994753f/contracts',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
     },
   >;
@@ -147,7 +147,7 @@ export default class TokenClient extends ContractClient {
     TokenClient,
     {
       function: 'totalSupply',
-      contract: 'erc20.sol',
+      contract: 'base.sol',
       // eslint-disable-next-line max-len
       contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
@@ -166,9 +166,9 @@ export default class TokenClient extends ContractClient {
     },
     TokenClient,
     {
-      contract: 'token.sol',
+      contract: 'Token.sol',
       // eslint-disable-next-line max-len
-      contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
+      contractPath: 'https://github.com/JoinColony/colonyToken/blob/7cc7d6b5bf3e94e6d97cd65583e3da38a994753f/contracts',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
     },
   >;
@@ -188,22 +188,6 @@ export default class TokenClient extends ContractClient {
       // eslint-disable-next-line max-len
       contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
-    },
-  >;
-  /*
-  Set the `name` of a token contract. This function can only be called by the current `owner` of the contract. In order to call token contract methods from within a colony, the token `owner` must be the address of the colony contract.
-  */
-  setName: TokenClient.Sender<
-    {
-      name: string, // The name of the token that will be set.
-    },
-    {},
-    TokenClient,
-    {
-      contract: 'token.sol',
-      // eslint-disable-next-line max-len
-      contractPath: 'https://github.com/dapphub/dappsys-monolithic/tree/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
-      version: '9ddf14118cd0436a19a7278282e4b8a607e14b54',
     },
   >;
   /*
@@ -235,7 +219,7 @@ export default class TokenClient extends ContractClient {
     {},
     TokenClient,
     {
-      contract: 'erc20.sol',
+      contract: 'base.sol',
       // eslint-disable-next-line max-len
       contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
@@ -255,9 +239,9 @@ export default class TokenClient extends ContractClient {
     },
     TokenClient,
     {
-      contract: 'Token.sol',
+      contract: 'base.sol',
       // eslint-disable-next-line max-len
-      contractPath: 'https://github.com/JoinColony/colonyToken/blob/2dc3136d71c0af9adb9b05e187885dd9dcae94c8/contracts',
+      contractPath: 'https://github.com/dapphub/dappsys-monolithic/blob/de9114c5fa1b881bf16b1414e7ed90cd3cb2e361',
       version: '8d3a50719cd51459db153006b5bd56c031e9d169',
     },
   >;
@@ -322,9 +306,6 @@ export default class TokenClient extends ContractClient {
     });
     this.addSender('burn', {
       input: [user, amount],
-    });
-    this.addSender('setName', {
-      input: [['name', 'bytes32String']],
     });
     this.addSender('setOwner', {
       input: [['owner', 'address']],