+++ CHANGELOG.md
@@ -139,6 +139,8 @@
 * Updated events in `ColonyNetworkClient` (`@colony/colony-js-client`)
   * Added `ColonyVersionAdded`
 
+* Updated `TokenClient` to reflect `colonyToken` contract (`@colony/colony-js-client`)
+
 ## v1.11.2
 
 **Bug Fixes**